Updated Beginner's Tour
As requested by SabreCat, I've done some editing on the Beginner's Tour. Mostly I just flesh out the guide to how the game works and try to contextualize everything on the page with the user avatar. If we implement these changes, there may need to be a few edits to the starting task descriptions to prevent overlap.

Definitely check to make sure I didn't mess up any of the coding. I'm a writer, not a programmer.