chore(changelog): add git-changelog config and sample CHANGELOG.md
This is the setup for the Grunt task that generates a changelog. Note
that it currently generates the changelog for the range
changelog-start..HEAD where changelog-start is a tag. It will always use
the most recent tag it can find. We may want to change this
behavior. Also, the original repo has some bugs, so I'm using my fork
until he accepts my PRs.

Migration notes: To run: `grunt git_changelog`