docs(rebirth): Bullet point about repurchase of limited ed gear after Rebirth
Requires commit 667a426 on habitrpg-shared. Adds a line to the Rebirth modal informing users that limited edition gear can be repurchased after Rebirth. Removed a line from the top of the modal to make room, preventing scrolling at ordinary resolutions--I'm open to suggestions if we think that line is crucial.