Fix #2879 (Check-Icons)
IE11 shows checkmark in wierd position. It seems that IE does size the "checkmark" differently than firefox, which make them appear to be more to the right and bottom. I'm using glyphicons to fix that. The look is constistent in different browser. Though its kind of thicker than the original was.