adjusted banned slurs - TRIGGER / CONTENT WARNING: slurs, swearwords, assault, etc
This deletes two swear words that were causing trouble for French
speakers (they mean "delay" or "late").
It also puts English variants of those words back to the list of slurs
(IIRC, previously the base words were in that list but were moved to
swears because of the French issue).